[spirv] Create separate files for GlPerVertex
